+++ crates/zng/src/env.rs
@@ -28,7 +28,7 @@
 //! Note that init **must be called in main**, it must be called in main to define the lifetime of the processes,
 //! this is needed to properly call [`on_process_exit`] handlers.
 //!
-//! Also see the [`init!`] section on WebAssembly for details on how to init in `"wasm32"` target builds.
+//! Also see [`init!`] docs for details init in `"wasm32"` and `"android"` target builds.
 //!
 //! # Full API
 //!